backend/requirements.txt

fastapi
uvicorn
python-dotenv
sqlmodel
requests
pydantic>=2

python-dateutil

fpdf2

python-telegram-bot==20.5
